% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_game}
\alias{analyze_game}
\title{Analyze a xiangqi game with a UCCI-compliant engine}
\usage{
analyze_game(moves, engine, multipv = 1)
}
\description{
Analyze a xiangqi game with a UCCI-compliant engine
}
