% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_moves.R
\name{clean_moves}
\alias{clean_moves}
\title{Parse a move string in a machine-readable format.}
\usage{
clean_moves(moves)
}
\arguments{
\item{moves}{chr vec of length 1}
}
\value{
vector of length n where n is the number of moves in the game.
}
\description{
UCCI engines require the move string to be clear of punctuation.
}
