% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%.\%}
\alias{\%.\%}
\title{Function to create operator that mimics an unary operator
Not meant to be called directly.}
\usage{
name \%.\% body
}
\arguments{
\item{name}{name of function to construct}

\item{body}{body of the function of operator to create}
}
\value{
seemingly unary infix function
}
\description{
Function to create operator that mimics an unary operator
Not meant to be called directly.
}
