% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_move.R
\name{position_move}
\alias{position_move}
\title{Create a xiangqi position}
\usage{
position_move(moves, pos = "start")
}
\arguments{
\item{moves}{vector length n, each element with uci(lan) notation, no whitespaces*}

\item{pos}{by default the starting position, otherwise named list of length 90.}
}
\description{

}
\details{
The notation system used in this position function is the one described by
the Universal Chinese Chess Protocol \strong{UCCI}, that which most xiangqi engines
use as the internal structure.
}
