% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_position.R
\name{print_position}
\alias{print_position}
\title{Print method to show the current position as a matrix in the console}
\usage{
print_position(pos = "startpos", style = "UCCI")
}
\arguments{
\item{pos}{Default is the starting position, otherwise a named list of length 90.}

\item{style}{one of \code{c("international", "traditional", "UCCI")}}
}
\value{
10*9 matrix depicting a xiangqi position.
}
\description{
see \code{position_move} for detailed information on how to construct positions.
}
