% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pgn.R
\name{read_pgn}
\alias{read_pgn}
\title{load a PGN file.}
\usage{
read_pgn(path, keep = TRUE)
}
\arguments{
\item{path}{filepath.}

\item{keep}{boolean. Should metadata be kept?}
}
\value{
list of length 2, 1 element with metadata, one with moves.
}
\description{
A convenience exists to read in multiple files, see \code{read_bulk}.
}
