% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pgn.R
\name{read_pgn_bulk}
\alias{read_pgn_bulk}
\title{Read a dense PGN file with manygames in a single file}
\usage{
read_pgn_bulk(path)
}
\arguments{
\item{path}{filepath}
}
\value{
list of length n games
}
\description{
Read a dense PGN file with manygames in a single file
}
