% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate move list to other notation systems}
\usage{
translate(moves, pos = "startpos")
}
\arguments{
\item{moves}{chr vec of n elements equal to n moves}

\item{pos}{list of length 90. See \code{position_move} to generate positions.}

\item{input_type}{one of \code{c("san", "lan")}}
}
\value{
chr vec of parsed notation, length 1.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

#' \tabular{lllll}{
1\tab \strong{pawn} \tab 兵 卒 \tab P\cr
2\tab \strong{canon} \tab 炮 \tab C\cr
3\tab \strong{rook} \tab 车 \tab R\cr
4\tab \strong{horse} \tab 马 \tab H\cr
5\tab \strong{elephant} \tab 相 象 \tab E\cr
5\tab \strong{advisor} \tab 仕 士 \tab A \cr
5\tab \strong{general} \tab 帅 将 \tab K \cr
}
}
\examples{
san <- c("炮二进二", "炮２进７", "炮八平二", "３进１")
lan <- c("h2h4", "b7b0", "b2h2", "f9e8")

translate(lan)
translate(san, input_type = "san")
}
